# public/api/public-openapi.yaml
openapi: 3.1.0
info:
  title: Changemaker Public API
  version: 1.1.0
  description: |
    The Changemaker public API exposes a small, security-hardened surface area
    for building integrations around user account management, invite
    redemption, and workspace awareness. All write operations are scoped to the
    authenticated Supabase session held by the caller.

    ### Authentication

    Most endpoints require a valid Supabase session. Sessions are issued by the
    Changemaker web application and conveyed via the `sb-access-token` and
    `sb-refresh-token` cookies. When developing locally, sign in through the web
    UI first so that the cookies are available in your browser or API client.

    ### Conventions

    * All responses are JSON encoded.
    * Dates are expressed in ISO-8601 format (UTC).
    * Rate limiting is enforced per-IP and/or per-resource. When a request is
      rejected due to throttling the response includes `retryAfter` in the body.
  contact:
    name: Changemaker Support
    email: support@changemaker.im
  termsOfService: https://changemaker.im/terms
servers:
  - url: https://changemaker.im
    description: Production
  - url: https://preview.changemaker.im
    description: Preview
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: |
        Send the Supabase session cookies issued after logging in (`sb-access-token`
        and `sb-refresh-token`). Include both values when making requests.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable description of the failure.
        code:
          type: string
          description: Application specific error code when available.
        retryAfter:
          type: integer
          format: int32
          description: Milliseconds to wait before retrying (present on 429 responses).
        details:
          type: object
          description: Additional context for validation errors.
          additionalProperties: true
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          description: Optional human-readable confirmation.
      required:
        - success
...
    EmailChangeConfirmRequest:
      type: object
      properties:
        token:
          type: string
          description: Verification token from the confirmation email.
      required:
        - token
...
    PasswordUpdateRequest:
      type: object
      description: Request body used when completing a Supabase password reset flow.
      properties:
        newPassword:
          type: string
          minLength: 8
      required:
        - newPassword
      additionalProperties: false
...
    WorkspaceSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        tenantId:
          type: string
        active:
          type: boolean
        published:
          type: boolean
      required:
        - id
        - name
        - slug
        - tenantId
        - active
        - published
...
    ChallengeSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        rewardType:
          type: string
          nullable: true
          enum: [points, sku, monetary]
      required:
        - id
        - title
        - description
        - status
        - startDate
        - endDate
    InviteAcceptRequest:
      type: object
      properties:
        code:
          type: string
          description: Invite code issued by a workspace admin.
      required:
        - code
    InviteAcceptResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        workspace:
          $ref: '#/components/schemas/WorkspaceSummary'
        challenge:
          allOf:
            - $ref: '#/components/schemas/ChallengeSummary'
            - nullable: true
        role:
          type: string
          description: Workspace role granted after accepting the invite.
          enum: [ADMIN, PARTICIPANT]
      required:
        - success
        - message
        - workspace
        - role
paths:
...
  /api/account/email/confirm:
    post:
      tags: [Account]
      operationId: confirmEmailChange
      summary: Confirm a pending email change
      description: Confirms a pending email change using the token supplied in the verification email.
      security:
        - SupabaseSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailChangeConfirmRequest'
            examples:
              confirm:
                value:
                  token: 9f8d1b47b7c34fb8b6a7b2e7c1f3a123
      responses:
        '200':
          description: Email successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                confirmed:
                  value:
                    success: true
                    message: Email successfully updated
        '400':
          description: Missing, invalid, or expired token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                expired:
                  value:
                    error: Token expired
        '401':
          description: Authentication failed or session missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/account/password/update:
    post:
      tags: [Password]
      operationId: completePasswordReset
      summary: Update password after reset verification
      description: |
        Completes a Supabase password reset flow after the user has verified
        their email via the reset link. Requires a valid Supabase session.
      security:
        - SupabaseSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordUpdateRequest'
            examples:
              updatePassword:
                value:
                  newPassword: 'AnotherStrongPass!2'
      responses:
        '200':
          description: Password updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                updated:
                  value:
                    success: true
                    message: Password updated successfully
        '400':
          description: Validation failed or Supabase rejected the update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rejected:
                  value:
                    error: Password must be at least 8 characters long
        '401':
          description: Authentication failed or session missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/invites/accept:
    post:
      tags: [Invites]
      operationId: acceptInviteCode
      summary: Redeem an invite code
      description: |
        Accepts an invite code and enrolls the current user into the target
        workspace (and optional challenge). A per-IP + code rate limit of **5
        attempts per minute** is enforced.
      security:
        - SupabaseSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteAcceptRequest'
            examples:
              redeem:
                value:
                  code: CMKR-8QF7-AB21
      responses:
        '200':
          description: Invite code redeemed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteAcceptResponse'
              examples:
                redeemed:
                  value:
                    success: true
                    message: Welcome to Acme Corp! You have been enrolled in the challenge.
                    workspace:
                      id: 13a4f884-2ed4-4e33-81c7-5ab5dd4da021
                      name: Acme Corp
                      slug: acme
                      tenantId: default
                      active: true
                      published: true
                    challenge:
                      id: 7a9b4a22-41e5-4bb3-8a20-46a3d0500d68
                      title: January Wellness Sprint
                      description: Complete daily wellness activities for 14 days.
                      status: PUBLISHED
                      startDate: '2025-01-15T00:00:00.000Z'
                      endDate: '2025-01-29T00:00:00.000Z'
                      rewardType: points
                    role: PARTICIPANT
        '400':
          description: Invite code missing, malformed, or not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing:
                  value:
                    error: Valid invite code is required
        '401':
          description: Authentication failed or session missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invite can only be redeemed by the intended recipient.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                mismatch:
                  value:
                    error: Invite code email mismatch
                    code: INVITE_EMAIL_MISMATCH
        '404':
          description: Invite code does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invite code exceeded the maximum allowed uses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                maxUses:
                  value:
                    error: Invite code usage limit reached
                    code: INVITE_MAX_USES
        '410':
          description: Invite code expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many attempts from the same IP/code combination.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                throttled:
                  value:
                    error: Too many attempts
                    code: RATE_LIMITED
                    retryAfter: 60000